# Conflicts:
#	src/main/java/it/safepet/backend/gestionePaziente/service/GestionePazienteServiceImpl.java

+++ src/main/java/it/safepet/backend/gestioneCartellaClinica/repository/VisitaMedicaRepository.java
@@ -3,5 +3,5 @@ package it.safepet.backend.gestioneCartellaClinica.repository;
 import it.safepet.backend.gestioneCartellaClinica.model.VisitaMedica;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VisitaMedicaRepository extends JpaRepository<VisitaMedica, Integer> {
+public interface VisitaMedicaRepository extends JpaRepository<VisitaMedica, Long> {
 }
